Move getThemeFilterTermsTable
